Convert bpAttributeWriter to bpAttributeWriteRead
